WIP: Code quality improvements.

+++ src/terminal/Color.h
@@ -239,7 +239,7 @@ struct ColorPalette
                     };
 
         // colors 232-255 are a grayscale ramp, intentionally leaving out black and white
-        for (uint8_t gray = 0, level = gray * 10 + 8; gray < 24; ++gray, level = gray * 10 + 8)
+        for (uint8_t gray = 0, level = static_cast<uint8_t>(gray * 10 + 8); gray < 24; ++gray, level = static_cast<uint8_t>(gray * 10 + 8))
             colors[232 + gray] = RGBColor{level, level, level};
 
         return colors;
